+++ src/plugins/blocksmith/components/article-block-modal/index.jsx
@@ -1,13 +1,12 @@
-import classNames from 'classnames';
-import inflection from 'inflection';
-import moment from 'moment';
-import PropTypes from 'prop-types';
-import React from 'react';
 import { connect } from 'react-redux';
 import ArticlesTable from '@triniti/cms/plugins/news/components/articles-table';
+import classNames from 'classnames';
 import createDelegateFactory from '@triniti/app/createDelegateFactory';
+import inflection from 'inflection';
 import Message from '@gdbots/pbj/Message';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
+import PropTypes from 'prop-types';
+import React from 'react';
 import SearchArticlesSort from '@triniti/schemas/triniti/news/enums/SearchArticlesSort';
 import {
   Modal,
@@ -19,10 +18,10 @@ import {
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
 import CustomizeOptions from './CustomizeOptions';
+import delegateFactory from './delegate';
 import Footer from './Footer';
 import Header from './Header';
 import SearchBar from '../search-bar';
-import delegateFactory from './delegate';
 import selector from './selector';
 
 const ARTICLE = 'article';
@@ -71,7 +70,7 @@ class ArticleBlockModal extends React.Component {
       selectedImageRef: imageRef || null,
       showImage: block.get('show_image'),
       aside: block.get('aside'),
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
     };
     this.handleAddBlock = this.handleAddBlock.bind(this);
     this.handleChangeCheckbox = this.handleChangeCheckbox.bind(this);
@@ -119,7 +118,7 @@ class ArticleBlockModal extends React.Component {
     return block.schema().createMessage()
       .set('node_ref', NodeRef.fromNode(selectedArticleNode))
       .set('show_image', !!showImage)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
+      .set('updated_date', hasUpdatedDate ? updatedDate : null)
       .set('image_ref', selectedImageRef || null)
       .set('link_text', linkText || null)
       .set('aside', aside);
@@ -146,7 +145,6 @@ class ArticleBlockModal extends React.Component {
   handleChangeQ({ target: { value: articleQ } }) {
     this.setState({ articleQ }, this.handleSearchArticles);
   }
-
 
   handleChangeTime({ target: { value: time } }) {
     this.setState(changedTime(time));
@@ -258,22 +256,19 @@ class ArticleBlockModal extends React.Component {
           toggle={toggle}
         />
         <ModalBody className="p-0">
-          {
-            activeStep === 0 && (
-              <SearchBar
-                onChangeQ={this.handleChangeQ}
-                onClick={this.handleSearchArticles}
-                placeholder={`Search ${inflection.pluralize(type)}...`}
-                value={articleQ}
-              />
-            )
-          }
+          {activeStep === 0 && (
+          <SearchBar
+            onChangeQ={this.handleChangeQ}
+            onClick={this.handleSearchArticles}
+            placeholder={`Search ${inflection.pluralize(type)}...`}
+            value={articleQ}
+          />
+          )}
           <ScrollableContainer
             className="bg-gray-400"
             style={{ height: `calc(100vh - ${activeStep === 0 ? 212 : 167}px)` }}
           >
-            {
-              activeStep === 0 && isReadyToDisplay && !!articleNodes.length
+            {activeStep === 0 && isReadyToDisplay && !!articleNodes.length
               && (
                 <ArticlesTable
                   hasCheckboxes={false}
@@ -289,10 +284,8 @@ class ArticleBlockModal extends React.Component {
                   sort={articleSort}
                   striped
                 />
-              )
-            }
-            {
-              activeStep === 1
+              )}
+            {activeStep === 1
               && (
                 <CustomizeOptions
                   aside={aside}
@@ -313,26 +306,21 @@ class ArticleBlockModal extends React.Component {
                   showImage={showImage}
                   updatedDate={updatedDate}
                 />
-              )
-            }
-            {
-              isReadyToDisplay && activeStep === 0 && !articleNodes.length
+              )}
+            {isReadyToDisplay && activeStep === 0 && !articleNodes.length
               && (
                 <div className="not-found-message">
                   <p>No articles found that match your search.</p>
                 </div>
-              )
-            }
-            {
-              !isReadyToDisplay && activeStep === 0
+              )}
+            {!isReadyToDisplay && activeStep === 0
               && (
                 <Spinner
                   centered
                   className={classNames({ 'mt-3': activeStep === 1 })}
                   style={activeStep === 1 ? { height: 'auto' } : {}}
                 />
-              )
-            }
+              )}
           </ScrollableContainer>
         </ModalBody>
         <Footer