+++ src/plugins/blocksmith/components/code-block-modal/index.jsx
@@ -1,26 +1,21 @@
-import moment from 'moment';
+import CodeBlockPreview from '@triniti/cms/plugins/blocksmith/components/code-block-preview';
+import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
+import Message from '@gdbots/pbj/Message';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Message from '@gdbots/pbj/Message';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 import {
   Button,
   Checkbox,
-  DatePicker,
   FormGroup,
   Icon,
   Input,
-  InputGroup,
-  InputGroupAddon,
-  InputGroupText,
   Label,
   Modal,
   ModalBody,
   ModalFooter,
   ModalHeader,
 } from '@triniti/admin-ui-plugin/components';
-import CodeBlockPreview from '@triniti/cms/plugins/blocksmith/components/code-block-preview';
-import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
-
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
@@ -45,7 +40,7 @@ export default class CodeBlockModal extends React.Component {
     this.state = {
       code: block.get('code'),
       hasUpdatedDate: block.has('updated_date'),
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       aside: block.get('aside'),
     };
     this.handleAddBlock = this.handleAddBlock.bind(this);
@@ -67,7 +62,7 @@ export default class CodeBlockModal extends React.Component {
     const { block } = this.props;
     return block.schema().createMessage()
       .set('code', code || null)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
+      .set('updated_date', hasUpdatedDate ? updatedDate : null)
       .set('aside', aside);
   }
 
@@ -126,39 +121,15 @@ export default class CodeBlockModal extends React.Component {
             <Icon imgSrc="info-outline" id="aside-tooltip" size="xs" className="ml-1" />
             <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
           </FormGroup>
-          {
-          hasUpdatedDate
+          {hasUpdatedDate
           && (
-          <FormGroup>
-            <Label>
-                  Updated Time: {updatedDate.format('YYYY-MM-DD hh:mm A')}
-            </Label>
-            <FormGroup className="mb-3 mt-1 shadow-none">
-              <DatePicker
-                onChange={this.handleChangeDate}
-                selected={updatedDate}
-                shouldCloseOnSelect={false}
-                inline
-              />
-              <InputGroup style={{ width: '15rem', margin: 'auto' }}>
-                <InputGroupAddon addonType="prepend" className="text-dark">
-                  <InputGroupText>
-                    <Icon imgSrc="clock-outline" />
-                  </InputGroupText>
-                </InputGroupAddon>
-                <Input
-                  type="time"
-                  onChange={this.handleChangeTime}
-                  defaultValue={updatedDate.format('HH:mm')}
-                />
-              </InputGroup>
-            </FormGroup>
-          </FormGroup>
-          )
-        }
-          {
-          code && <CodeBlockPreview block={this.setBlock()} />
-        }
+            <DateTimePicker
+              onChangeDate={this.handleChangeDate}
+              onChangeTime={this.handleChangeTime}
+              updatedDate={updatedDate}
+            />
+          )}
+          {code && <CodeBlockPreview block={this.setBlock()} />}
         </ModalBody>
         <ModalFooter>
           <Button onClick={toggle}>Cancel</Button>