+++ src/plugins/blocksmith/components/google-map-block-modal/index.jsx
@@ -1,13 +1,12 @@
-import moment from 'moment';
+import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
+import GoogleMapBlockPreview from '@triniti/cms/plugins/blocksmith/components/google-map-block-preview';
+import Message from '@gdbots/pbj/Message';
 import PropTypes from 'prop-types';
 import React from 'react';
-import Message from '@gdbots/pbj/Message';
-
-import GoogleMapBlockPreview from '@triniti/cms/plugins/blocksmith/components/google-map-block-preview';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 import {
   Button,
   Checkbox,
-  DatePicker,
   Dropdown,
   DropdownItem,
   DropdownMenu,
@@ -15,16 +14,12 @@ import {
   FormGroup,
   Icon,
   Input,
-  InputGroup,
-  InputGroupAddon,
-  InputGroupText,
   Label,
   Modal,
   ModalBody,
   ModalFooter,
   ModalHeader,
 } from '@triniti/admin-ui-plugin/components';
-import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
@@ -58,7 +53,7 @@ export default class GoogleMapBlockModal extends React.Component {
       mapType: block.get('maptype'),
       q: block.get('q') || '',
       touched: false,
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       zoom: block.get('zoom'),
     };
     this.handleAddBlock = this.handleAddBlock.bind(this);
@@ -83,11 +78,11 @@ export default class GoogleMapBlockModal extends React.Component {
     const { aside, hasUpdatedDate, mapType, q, updatedDate, zoom } = this.state;
     const { block } = this.props;
     return block.schema().createMessage()
+      .set('aside', aside)
       .set('maptype', mapType || null)
       .set('q', q || null)
-      .set('zoom', zoom || null)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
-      .set('aside', aside);
+      .set('updated_date', hasUpdatedDate ? updatedDate : null)
+      .set('zoom', zoom || null);
   }
 
   handleAddBlock() {
@@ -200,8 +195,7 @@ export default class GoogleMapBlockModal extends React.Component {
               Auto Zoom
             </Checkbox>
           </FormGroup>
-          {
-            !isAutoZoom
+          {!isAutoZoom
             && (
               <FormGroup>
                 <Label>Zoom</Label>
@@ -211,8 +205,7 @@ export default class GoogleMapBlockModal extends React.Component {
                   value={zoom}
                 />
               </FormGroup>
-            )
-          }
+            )}
           <FormGroup>
             <Label>Map type</Label>
             <Dropdown isOpen={isDropdownOpen} toggle={this.handleToggleDropDown}>
@@ -220,18 +213,16 @@ export default class GoogleMapBlockModal extends React.Component {
                 {mapType}
               </DropdownToggle>
               <DropdownMenu>
-                {
-                  mapTypes
-                    .filter((type) => type !== mapType)
-                    .map((type) => (
-                      <DropdownItem
-                        key={type}
-                        onClick={this.handleChangeMapType}
-                      >
-                        {type}
-                      </DropdownItem>
-                    ))
-                }
+                {mapTypes
+                  .filter((type) => type !== mapType)
+                  .map((type) => (
+                    <DropdownItem
+                      key={type}
+                      onClick={this.handleChangeMapType}
+                    >
+                      {type}
+                    </DropdownItem>
+                  ))}
               </DropdownMenu>
             </Dropdown>
           </FormGroup>
@@ -245,51 +236,27 @@ export default class GoogleMapBlockModal extends React.Component {
             <Icon imgSrc="info-outline" id="aside-tooltip" size="xs" className="ml-1" />
             <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
           </FormGroup>
-          {
-            hasUpdatedDate
+          {hasUpdatedDate
             && (
               <div className="modal-body-blocksmith">
-                <FormGroup>
-                  <Label>
-                    Updated Time: {updatedDate.format('YYYY-MM-DD hh:mm A')}
-                  </Label>
-                  <FormGroup className="mb-3 mt-1 shadow-none">
-                    <DatePicker
-                      onChange={this.handleChangeDate}
-                      selected={updatedDate}
-                      shouldCloseOnSelect={false}
-                      inline
-                    />
-                    <InputGroup style={{ width: '15rem', margin: 'auto' }}>
-                      <InputGroupAddon addonType="prepend" className="text-dark">
-                        <InputGroupText>
-                          <Icon imgSrc="clock-outline" />
-                        </InputGroupText>
-                      </InputGroupAddon>
-                      <Input
-                        type="time"
-                        onChange={this.handleChangeTime}
-                        defaultValue={updatedDate.format('HH:mm')}
-                      />
-                    </InputGroup>
-                  </FormGroup>
-                </FormGroup>
+                <DateTimePicker
+                  onChangeDate={this.handleChangeDate}
+                  onChangeTime={this.handleChangeTime}
+                  updatedDate={updatedDate}
+                />
               </div>
-            )
-          }
+            )}
           {
             !isValid && touched
             && <p className="text-danger">{errorMsg}</p>
           }
-          {
-            isValid
+          {isValid
             && (
               <GoogleMapBlockPreview
                 block={this.setBlock()}
                 width={528}
               />
-            )
-          }
+            )}
         </ModalBody>
         <ModalFooter>
           <Button onClick={toggle}>Cancel</Button>