+++ src/plugins/blocksmith/components/poll-block-modal/CustomizeOptions.jsx
@@ -1,20 +1,14 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import moment from 'moment';
+import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
 import Message from '@gdbots/pbj/Message';
+import PollBlockPreview from '@triniti/cms/plugins/blocksmith/components/poll-block-preview';
+import PropTypes from 'prop-types';
+import React from 'react';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 import {
   Checkbox,
-  DatePicker,
   FormGroup,
   Icon,
-  Input,
-  InputGroup,
-  InputGroupAddon,
-  InputGroupText,
-  Label,
 } from '@triniti/admin-ui-plugin/components';
-import PollBlockPreview from '@triniti/cms/plugins/blocksmith/components/poll-block-preview';
-import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 
 const CustomizeOptions = ({
   aside,
@@ -39,36 +33,14 @@ const CustomizeOptions = ({
         <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
       </FormGroup>
     </FormGroup>
-    {
-      hasUpdatedDate
-      && (
-        <FormGroup>
-          <Label>
-            Updated Time: {updatedDate.format('YYYY-MM-DD hh:mm A')}
-          </Label>
-          <FormGroup className="mb-3 mt-1 shadow-none">
-            <DatePicker
-              onChange={handleChangeDate}
-              selected={updatedDate}
-              shouldCloseOnSelect={false}
-              inline
-            />
-            <InputGroup style={{ width: '15rem', margin: 'auto' }}>
-              <InputGroupAddon addonType="prepend" className="text-dark">
-                <InputGroupText>
-                  <Icon imgSrc="clock-outline" />
-                </InputGroupText>
-              </InputGroupAddon>
-              <Input
-                type="time"
-                onChange={handleChangeTime}
-                defaultValue={updatedDate.format('HH:mm')}
-              />
-            </InputGroup>
-          </FormGroup>
-        </FormGroup>
-      )
-    }
+    {hasUpdatedDate
+    && (
+      <DateTimePicker
+        onChangeDate={handleChangeDate}
+        onChangeTime={handleChangeTime}
+        updatedDate={updatedDate}
+      />
+    )}
   </div>
 );
 
@@ -79,7 +51,7 @@ CustomizeOptions.propTypes = {
   onChangeCheckBox: PropTypes.func.isRequired,
   onChangeDate: PropTypes.func.isRequired,
   onChangeTime: PropTypes.func.isRequired,
-  updatedDate: PropTypes.instanceOf(moment).isRequired,
+  updatedDate: PropTypes.instanceOf(Date).isRequired,
 };
 
 export default CustomizeOptions;