+++ src/plugins/blocksmith/components/twitter-tweet-block-modal/index.jsx
@@ -1,10 +1,9 @@
-import moment from 'moment';
-import React from 'react';
-import PropTypes from 'prop-types';
-import Message from '@gdbots/pbj/Message';
 import { Button, Checkbox, FormGroup, Input, Label, Modal, ModalHeader, ModalBody, ModalFooter, Icon, UncontrolledTooltip } from '@triniti/admin-ui-plugin/components';
-import TwitterTweetBlockPreview from '@triniti/cms/plugins/blocksmith/components/twitter-tweet-block-preview';
 import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
+import Message from '@gdbots/pbj/Message';
+import PropTypes from 'prop-types';
+import React from 'react';
+import TwitterTweetBlockPreview from '@triniti/cms/plugins/blocksmith/components/twitter-tweet-block-preview';
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
@@ -42,7 +41,7 @@ export default class TwitterTweetBlockModal extends React.Component {
       touched: false,
       tweetId,
       tweetText: block.get('tweet_text'),
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       url: tweetId && screenName ? `https://twitter.com/${screenName}/status/${tweetId}` : '',
     };
     this.handleAddBlock = this.handleAddBlock.bind(this);
@@ -70,13 +69,13 @@ export default class TwitterTweetBlockModal extends React.Component {
     } = this.state;
     const { block } = this.props;
     return block.schema().createMessage()
+      .set('aside', aside)
       .set('hide_media', hideMedia)
       .set('hide_thread', hideThread)
       .set('screen_name', screenName || null)
       .set('tweet_id', tweetId || null)
       .set('tweet_text', tweetText || null)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
-      .set('aside', aside);
+      .set('updated_date', hasUpdatedDate ? updatedDate : null);
   }
 
   handleAddBlock() {
@@ -188,8 +187,7 @@ export default class TwitterTweetBlockModal extends React.Component {
             <Icon imgSrc="info-outline" id="aside-tooltip" size="xs" className="ml-1" />
             <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
           </FormGroup>
-          {
-            hasUpdatedDate
+          {hasUpdatedDate
             && (
               <div className="modal-body-blocksmith">
                 <DateTimePicker
@@ -198,8 +196,7 @@ export default class TwitterTweetBlockModal extends React.Component {
                   updatedDate={updatedDate}
                 />
               </div>
-            )
-          }
+            )}
           {
             isValid && <TwitterTweetBlockPreview block={this.setBlock()} />
           }