+++ src/plugins/blocksmith/components/video-block-modal/index.jsx
@@ -1,10 +1,9 @@
-import moment from 'moment';
-import PropTypes from 'prop-types';
-import React from 'react';
 import { connect } from 'react-redux';
-
 import createDelegateFactory from '@triniti/app/createDelegateFactory';
 import Message from '@gdbots/pbj/Message';
+import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
+import PropTypes from 'prop-types';
+import React from 'react';
 import VideosTable from '@triniti/cms/plugins/ovp/components/videos-table';
 import {
   Modal,
@@ -12,15 +11,14 @@ import {
   ScrollableContainer,
   Spinner,
 } from '@triniti/admin-ui-plugin/components';
-import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
 import CustomizeOptions from './CustomizeOptions';
+import delegateFactory from './delegate';
 import Footer from './Footer';
 import Header from './Header';
 import SearchBar from '../search-bar';
-import delegateFactory from './delegate';
 import selector from './selector';
 import './styles.scss';
 
@@ -57,18 +55,18 @@ class VideoBlockModal extends React.Component {
     const { block, imageNode, video } = props;
     this.state = {
       activeStep: 0,
+      aside: block.get('aside'),
       autoplay: block.get('autoplay'),
-      launchText: block.get('launch_text'),
       hasLaunchText: block.has('launch_text'),
       hasUpdatedDate: block.has('updated_date'),
       isAssetPickerModalOpen: false,
       isMuted: block.get('muted'),
+      launchText: block.get('launch_text'),
       q: '',
       selectedImageNode: imageNode || null,
       selectedVideo: video || null,
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       willShowMoreVideos: block.get('show_more_videos'),
-      aside: block.get('aside'),
     };
     this.handleAddBlock = this.handleAddBlock.bind(this);
     this.handleChangeCheckbox = this.handleChangeCheckbox.bind(this);
@@ -110,14 +108,14 @@ class VideoBlockModal extends React.Component {
     } = this.state;
     const { block } = this.props;
     return block.schema().createMessage()
+      .set('aside', aside)
       .set('autoplay', autoplay)
       .set('launch_text', (hasLaunchText && launchText) || null)
       .set('muted', isMuted)
       .set('node_ref', selectedVideo.get('_id').toNodeRef())
-      .set('show_more_videos', willShowMoreVideos)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
       .set('poster_image_ref', selectedImageNode ? NodeRef.fromNode(selectedImageNode) : null)
-      .set('aside', aside);
+      .set('show_more_videos', willShowMoreVideos)
+      .set('updated_date', hasUpdatedDate ? updatedDate : null);
   }
 
   handleAddBlock() {
@@ -244,8 +242,7 @@ class VideoBlockModal extends React.Component {
           toggle={toggle}
         />
         <ModalBody className="p-0 bg-gray-400">
-          {
-            activeStep === 0
+          {activeStep === 0
             && (
             <SearchBar
               onChangeQ={this.handleChangeQ}
@@ -253,8 +250,7 @@ class VideoBlockModal extends React.Component {
               placeholder="search videos..."
               value={q}
             />
-            )
-          }
+            )}
           <ScrollableContainer
             className="bg-gray-400"
             style={{ height: `calc(100vh - ${activeStep === 0 ? 212 : 167}px)` }}
@@ -263,8 +259,7 @@ class VideoBlockModal extends React.Component {
             {
               !isFulfilled && <Spinner centered />
             }
-            {
-              activeStep === 0 && isFulfilled
+            {activeStep === 0 && isFulfilled
               && (
                 <VideosTable
                   hasCheckboxes={false}
@@ -280,10 +275,8 @@ class VideoBlockModal extends React.Component {
                   sort={sort}
                   striped
                 />
-              )
-            }
-            {
-              activeStep === 1
+              )}
+            {activeStep === 1
               && (
                 <CustomizeOptions
                   autoplay={autoplay}
@@ -306,8 +299,7 @@ class VideoBlockModal extends React.Component {
                   willShowMoreVideos={willShowMoreVideos}
                   aside={aside}
                 />
-              )
-            }
+              )}
           </ScrollableContainer>
         </ModalBody>
         <Footer