+++ src/plugins/blocksmith/components/vimeo-video-block-modal/index.jsx
@@ -1,19 +1,17 @@
-import moment from 'moment';
-import React from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import Message from '@gdbots/pbj/Message';
-import he from 'he';
 import { Button, Checkbox, FormGroup, Icon, Input, Label, Modal, ModalHeader, ModalBody, ModalFooter } from '@triniti/admin-ui-plugin/components';
-import VimeoVideoBlockPreview from '@triniti/cms/plugins/blocksmith/components/vimeo-video-block-preview';
+import { connect } from 'react-redux';
 import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
+import he from 'he';
 import ImageAssetPicker from '@triniti/cms/plugins/dam/components/image-asset-picker';
-import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
+import Message from '@gdbots/pbj/Message';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
+import PropTypes from 'prop-types';
+import React from 'react';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
+import VimeoVideoBlockPreview from '@triniti/cms/plugins/blocksmith/components/vimeo-video-block-preview';
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
-
 import selector from './selector';
 
 const EMBED_REGEX = /https?:\/\/player\.vimeo\.com\/video\/\d+/;
@@ -71,7 +69,7 @@ class VimeoVideoBlockModal extends React.Component {
       selectedImageNode: imageNode || null,
       title: block.get('title'),
       touched: false,
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       userId: block.get('user_id'),
       userName: block.get('user_name'),
       willLoop: block.get('loop'),
@@ -109,19 +107,19 @@ class VimeoVideoBlockModal extends React.Component {
     } = this.state;
     const { block } = this.props;
     return block.schema().createMessage()
-      .set('id', id || null)
+      .set('aside', aside)
       .set('autoplay', autoplay)
+      .set('description', description || null)
+      .set('id', id || null)
       .set('loop', willLoop)
+      .set('poster_image_ref', selectedImageNode ? NodeRef.fromNode(selectedImageNode) : null)
       .set('show_byline', willShowByline)
       .set('show_portrait', willShowPortrait)
       .set('show_title', willShowTitle)
-      .set('description', description || null)
       .set('title', title || null)
+      .set('updated_date', hasUpdatedDate ? updatedDate : null)
       .set('user_id', userId || null)
-      .set('user_name', userName || null)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
-      .set('poster_image_ref', selectedImageNode ? NodeRef.fromNode(selectedImageNode) : null)
-      .set('aside', aside);
+      .set('user_name', userName || null);
   }
 
   handleAddBlock() {
@@ -398,8 +396,7 @@ class VimeoVideoBlockModal extends React.Component {
             <Icon imgSrc="info-outline" id="aside-tooltip" size="xs" className="ml-1" />
             <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
           </FormGroup>
-          {
-            hasUpdatedDate
+          {hasUpdatedDate
             && (
               <div className="modal-body-blocksmith">
                 <DateTimePicker
@@ -408,8 +405,7 @@ class VimeoVideoBlockModal extends React.Component {
                   updatedDate={updatedDate}
                 />
               </div>
-            )
-          }
+            )}
         </ModalBody>
         <ModalFooter>
           <Button onClick={toggle} innerRef={(el) => { this.button = el; }}>Cancel</Button>