+++ src/plugins/blocksmith/components/youtube-video-block-modal/index.jsx
@@ -1,9 +1,13 @@
-import moment from 'moment';
-import React from 'react';
-import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-
+import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
+import ImageAssetPicker from '@triniti/cms/plugins/dam/components/image-asset-picker';
 import Message from '@gdbots/pbj/Message';
+import moment from 'moment';
+import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
+import PropTypes from 'prop-types';
+import React from 'react';
+import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
+import YoutubeVideoBlockPreview from '@triniti/cms/plugins/blocksmith/components/youtube-video-block-preview';
 import {
   Button,
   Checkbox,
@@ -16,11 +20,6 @@ import {
   ModalBody,
   ModalFooter,
 } from '@triniti/admin-ui-plugin/components';
-import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
-import YoutubeVideoBlockPreview from '@triniti/cms/plugins/blocksmith/components/youtube-video-block-preview';
-import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
-import ImageAssetPicker from '@triniti/cms/plugins/dam/components/image-asset-picker';
-import UncontrolledTooltip from '@triniti/cms/plugins/common/components/uncontrolled-tooltip';
 
 import changedDate from '../../utils/changedDate';
 import changedTime from '../../utils/changedTime';
@@ -65,7 +64,7 @@ class YouTubeVideoBlockModal extends React.Component {
       selectedImageNode: imageNode || null,
       startAt: block.get('start_at'),
       touched: false,
-      updatedDate: block.has('updated_date') ? moment(block.get('updated_date')) : moment(),
+      updatedDate: block.get('updated_date', new Date()),
       url: id ? `https://www.youtube.com/watch?v=${id}` : '',
     };
 
@@ -94,12 +93,12 @@ class YouTubeVideoBlockModal extends React.Component {
     } = this.state;
     const { block } = this.props;
     return block.schema().createMessage()
+      .set('aside', aside)
       .set('autoplay', autoplay)
       .set('id', id || null)
-      .set('start_at', startAt || null)
-      .set('updated_date', hasUpdatedDate ? updatedDate.toDate() : null)
       .set('poster_image_ref', selectedImageNode ? NodeRef.fromNode(selectedImageNode) : null)
-      .set('aside', aside);
+      .set('start_at', startAt || null)
+      .set('updated_date', hasUpdatedDate ? updatedDate : null);
   }
 
   handleAddBlock() {
@@ -267,16 +266,14 @@ class YouTubeVideoBlockModal extends React.Component {
               && <p>Will start at: {moment.utc(startAt * 1000).format('HH:mm:ss')}</p>
             }
           </FormGroup>
-          {
-            isValid && id
+          {isValid && id
             && (
               <YoutubeVideoBlockPreview
                 block={this.setBlock()}
                 className="mb-3"
                 width={526}
               />
-            )
-          }
+            )}
           <FormGroup>
             <ImageAssetPicker
               multiAssetErrorMessage="Invalid Action: Trying to assign multiple YouTube Block Poster images."
@@ -301,14 +298,13 @@ class YouTubeVideoBlockModal extends React.Component {
             </Checkbox>
           </FormGroup>
           <FormGroup>
-            <Checkbox size="sd" id ="aside" checked={aside} onChange={this.handleChangeCheckbox}>
+            <Checkbox size="sd" id="aside" checked={aside} onChange={this.handleChangeCheckbox}>
               Aside
             </Checkbox>
             <Icon imgSrc="info-outline" id="aside-tooltip" size="xs" className="ml-1" />
             <UncontrolledTooltip target="aside-tooltip">Is only indirectly related to the main content.</UncontrolledTooltip>
           </FormGroup>
-          {
-            hasUpdatedDate
+          {hasUpdatedDate
             && (
               <div className="modal-body-blocksmith">
                 <DateTimePicker
@@ -317,8 +313,7 @@ class YouTubeVideoBlockModal extends React.Component {
                   updatedDate={updatedDate}
                 />
               </div>
-            )
-          }
+            )}
         </ModalBody>
         <ModalFooter>
           <Button onClick={toggle} innerRef={(el) => { this.button = el; }}>Cancel</Button>